backend/requirements.txt
fastapi
uvicorn
pymongo
pandas
genson
python-dotenv
